feat(aluguel): adiciona lógica de cálculo para data de entrega, devolução e valor total

+++ locadoraVeiculos/src/main/java/com/solutis/locadoraVeiculos/dtos/carroDtos/LerCarroDto.java
@@ -1,15 +1,13 @@
 package com.solutis.locadoraVeiculos.dtos.carroDtos;
 
-import com.solutis.locadoraVeiculos.model.Carro;
-import java.math.BigDecimal;
-import java.util.List;
-
-import io.swagger.v3.oas.annotations.media.Schema;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.solutis.locadoraVeiculos.model.Acessorio;
+import com.solutis.locadoraVeiculos.model.Carro;
 import com.solutis.locadoraVeiculos.model.ModeloCarro;
-
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+
+import java.util.List;
 
 
 @Data
@@ -37,7 +35,7 @@ public class LerCarroDto {
 
     private String cor;
 
-    private BigDecimal valorDiaria;
+    private Double valorDiaria;
 
     @JsonIgnoreProperties({"carros"})
     private List<Acessorio> acessorios;
